Smaller Rust inverted index (#7187)
* u16 number of block entries

This matches the C definition but the size of an `IndexBlock` does not
change because of 8 byte alignment.

* u32 for number of index docs

This aligns with the C definition and reduces the size of the inverted index.

* u32 for GC marker

This aligns with the C definition. Memory is not reduced because of the
Rust 8 byte alignment.

* Check for overflow

* Update missed headers

* Upcast size in iterator

* Update iterator test

+++ src/redisearch_rs/rqe_iterators/src/inverted_index.rs
@@ -109,7 +109,7 @@ where
     }
 
     fn num_estimated(&self) -> usize {
-        self.reader.unique_docs()
+        self.reader.unique_docs() as usize
     }
 
     fn last_doc_id(&self) -> t_docId {
